fix styling and logic in UserPanel, UpdateUserModal - not finshed
