+++ src/components/CustomButton/CustomButton.jsx
@@ -6,7 +6,7 @@ import css from './CustomButton.module.css';
 
 const CustomButton = ({ type, color, content, onClick = null }) => {
   const btnClass = color =>
-    `${ color === 'primary' ? css.buttonPrimary : color === 'secondary' ? css.buttonSecondary : null }`;
+    `${ color === 'primary' ? css.buttonPrimary : color === 'secondary' ? css.buttonSecondary : color === 'red' ? css.buttonRed : null }`;
   return (
     <button type={type} onClick={onClick} className={btnClass(color)}>
       {content}
