+++ src/components/Header/Header.module.css
@@ -17,8 +17,10 @@
   display: flex;
   justify-content: space-between;
   align-items: center;
+  align-content: center;
   height: 100%;
   padding-top: 0px;
+  margin: auto;
 }
 
 .logoWrapper {
@@ -32,6 +34,7 @@
   justify-content: right;
   align-items: center;
   gap: 20px;
+  position: relative;
 }
 
 .exitWrapper {
@@ -60,7 +63,10 @@
   font-family: var(--font-secondary);
   font-size: 18px;
   font-weight: 400;
+  padding-block: 0;
+  margin-block: 0;
   margin-right: 8px;
+  
   @media screen and (min-width: 768px) {
     margin-right: 12px;
   }
@@ -68,6 +74,7 @@
 
 .exit {
   display: none;
+  
   @media screen and (min-width: 768px) {
     display: block;
     color: var(--color-text-nav-menu-gray);
@@ -78,6 +85,87 @@
   }
 }
  
-.text {
+.btn1 {
+  position: relative;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 5px;
+  width: 38px;
+  height: 40px;
+  transition-duration: 0.5s;
+}
+.btn1 .icon {
+  transition-duration: 0.5s;
+  position: absolute;
+  height: 4px;
+  width: 30px;
+  top: 18px;
+  left: 4px;
+  background-color: var(--color-text-nav-menu-gray);
+}
+.btn1 .icon:before {
+  transition-duration: 0.5s;
+  position: absolute;
+  width: 30px;
+  height: 4px;
+  background-color: var(--color-text-nav-menu-gray);
+  content: "";
+  top: -10px;
+  left: 0px;
+}
+.btn1 .icon:after {
+  transition-duration: 0.5s;
+  position: absolute;
+  width: 30px;
+  height: 4px;
+  background-color: var(--color-text-nav-menu-gray);
+  content: "";
+  top: 10px;
+  left: 0px;
+}
+.btn1open {
+  position: relative;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 5px;
+  width: 38px;
+  height: 40px;
+  transition-duration: 0.5s;
+}
+.btn1open .icon {
+  transition-duration: 0.5s;
+  position: absolute;
+  height: 4px;
+  width: 30px;
+  top: 18px;
+  left: 3px;
+  background: transparent;
+}
+.btn1open .icon:before {
+  transition-duration: 0.5s;
+  position: absolute;
+  width: 30px;
+  height: 4px;
+  background-color: var(--color-text-nav-menu-gray);
+  content: "";
+  top: -10px;
+  left: 0px;
+  transform: rotateZ(45deg) scaleX(1.25) translate(6.5px, 6.5px);
+}
+.btn1open .icon:after {
+  transition-duration: 0.5s;
+  position: absolute;
+  width: 30px;
+  height: 4px;
+  background-color: var(--color-text-nav-menu-gray);
+  content: "";
+  top: 10px;
+  left: 0px;
+  transform: rotateZ(-45deg) scaleX(1.25) translate(6px, -6px);
+}
+.btn1:hover,
+.btn1open:hover {
   cursor: pointer;
 }
