+++ src/components/UserPanel/UpdateUserModal/UpdateUserModal.module.css
@@ -1,0 +1,71 @@
+/* .container_input {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  padding: 10px;
+}
+
+.form {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  border-radius: 20px;
+  padding: 0;
+  margin: 0;
+  width: 100vw;
+  max-width: 450px;
+  height: 100vh;
+
+  @media screen and (min-width: 768px) {
+    padding: 50px;
+  }
+}
+
+.container_form {
+  border-radius: 20px;
+}
+
+.button_container {
+  padding: 0;
+  margin: 0 auto;
+} */
+
+.form {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  background-color: var(--color-bg-secondary);
+  box-shadow: 0px 10px 60px rgba(170, 178, 197, 0.2);
+
+  @media screen and (min-width: 768px) {
+    width: 500px;
+    height: fit-content;
+    padding: 40px 0 50px;
+    border-radius: 20px;
+  }
+}
+
+.form-field {
+  position: relative;
+  margin-bottom: 40px;
+}
+
+.container_input {
+  position: relative;
+}
+
+.button_container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  margin: 0px;
+  padding: 0px;
+  gap: 20px;
+}
+
