+++ src/components/UserPanel/UserPanel.module.css
@@ -64,7 +64,7 @@
   }
   .paragraph {
     width: 100%;
-    padding: 5px;
+    padding: 0;
     margin: 5px auto;
   }
