--- src/components/UserPanel/UserPanelModal.module.css
@@ -1,13 +1,0 @@
-.container_input {
-    display: flex;
-    justify-content: center;
-    /* margin-top: 20px; */
-  }
-
-  .form {
-    border-radius: 20px; 
-  }  
-
-  .container_form {
-    border-radius: 20px;
-  }
